requirements.txt
accelerate==0.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==25.3.0
av==15.0.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bleach==6.1.0
brotli==1.1.0
cffi==1.16.0
charset-normalizer==3.4.3
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
diffusers==0.35.1
docopt==0.6.2
facexlib==0.3.0
fastapi==0.116.1
fastjsonschema==2.21.2
ffmpy==0.6.1
filelock==3.19.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2025.3.0
future==1.0.0
gfpgan==1.3.8
GPUtil==1.4.0
gradio==5.44.0
gradio_client==1.12.1
#h11==0.14.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.12.3
Jinja2==3.1.6
jedi==0.19.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy-loader==0.4
lmdb==1.7.3
llvmlite>=0.44.0,<0.45
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
optimum==1.27.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.0
pefile==2023.2.7
pickleshare==0.7.5
Pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
#python-multipart==0.0.9
pytz==2024.1
PySide6==6.7.3
PyYAML==6.0.2
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
realesrgan @ git+https://github.com/xinntao/Real-ESRGAN.git@a4abfb2979a7bbff3f69f58f58ae324608821e27
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rpds-py==0.18.0
huggingface-hub>=0.34.0,<1.0

ruff==0.12.11
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.47.3
sympy==1.13.1

tensorboard==2.16.2
tifffile==2025.5.10
tokenizers==0.21.4
tomli==2.0.1
tomlkit==0.12.4
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchsde==0.2.6
torchvision==0.21.0+cu124
trampoline==0.1.2
traitlets==5.14.2
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
vulture==2.14
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarg==0.1.9
#yarl==1.9.4
zipp==3.23.0

# ---- NOTE ----
# Torch CUDA wheels require:
#   pip install --index-url https://download.pytorch.org/whl/cu124 torch==2.6.0+cu124 torchvision==0.21.0+cu124 torchaudio==2.6.0+cu124
